+++ stroom-index-server/src/main/java/stroom/search/server/RemoteSearchResultFactory.java
@@ -6,14 +6,19 @@ import stroom.query.common.v2.Payload;
 import stroom.search.coprocessor.Coprocessors;
 import stroom.search.resultsender.NodeResult;
 import stroom.task.server.TaskContext;
+import stroom.util.logging.LambdaLogger;
+import stroom.util.logging.LambdaLoggerFactory;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.TimeUnit;
 
 public class RemoteSearchResultFactory {
+    private static final LambdaLogger LOGGER = LambdaLoggerFactory.getLogger(RemoteSearchResultFactory.class);
+
     private volatile Coprocessors coprocessors;
     private volatile CompletionState searchComplete;
     private volatile LinkedBlockingQueue<String> errors;
@@ -22,34 +27,36 @@ public class RemoteSearchResultFactory {
     private volatile boolean started;
 
     public NodeResult create() {
-        if (!started) {
-            return new NodeResult(Collections.emptyMap(), Collections.emptyList(), false);
-        } else if (taskContext.isTerminated()) {
+        try {
+            if (!started) {
+                return new NodeResult(Collections.emptyMap(), Collections.emptyList(), false);
+            } else if (taskContext.isTerminated()) {
+                return new NodeResult(null, null, true);
+            }
+
+            // Wait to complete.
+            final boolean complete = searchComplete.awaitCompletion(1, TimeUnit.SECONDS);
+
+            // Produce payloads for each coprocessor.
+            final Map<CoprocessorSettingsMap.CoprocessorKey, Payload> payloadMap = coprocessors.createPayloads();
+
+            // Drain all current errors to a list.
+            List<String> errorsSnapshot = new ArrayList<>();
+            errors.drainTo(errorsSnapshot);
+            if (errorsSnapshot.size() == 0) {
+                errorsSnapshot = null;
+            }
+
+            // Form a result to send back to the requesting node.
+            return new NodeResult(payloadMap, errorsSnapshot, complete);
+
+        } catch (final InterruptedException e) {
+            LOGGER.debug(e::getMessage, e);
+            // Keep interrupting.
+            Thread.currentThread().interrupt();
+
             return new NodeResult(null, null, true);
         }
-
-//        taskContext.setName("Search Result Sender");
-//        taskContext.info("Creating search result");
-
-        // Find out if searching is complete.
-        final boolean complete = searchComplete.isComplete();
-
-        // Produce payloads for each coprocessor.
-        final Map<CoprocessorSettingsMap.CoprocessorKey, Payload> payloadMap = coprocessors.createPayloads();
-
-        // Drain all current errors to a list.
-        List<String> errorsSnapshot = new ArrayList<>();
-        errors.drainTo(errorsSnapshot);
-        if (errorsSnapshot.size() == 0) {
-            errorsSnapshot = null;
-        }
-
-        // Form a result to send back to the requesting node.
-        final NodeResult result = new NodeResult(payloadMap, errorsSnapshot, complete);
-
-//        // Give the result to the callback.
-//        taskContext.info("Delivering search result");
-        return result;
     }
 
     public synchronized void destroy() {